Issue #683 Move builder to proper package
